% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{as.vector.pair}
\alias{as.vector.pair}
\title{pair's as.vector method}
\usage{
as.vector.pair(x, mode)
}
\arguments{
\item{x}{the object of class pair to convert}

\item{mode}{integer, character, etc...}
}
\description{
pair's as.vector method
}

