% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{const}
\alias{const}
\title{creates a constant function}
\usage{
const(v = stop("v not supplied"))
}
\arguments{
\item{v}{the return value of the constant function}
}
\value{
\code{function() { return (v) }}
}
\description{
creates a constant function
}
\details{
This is sugar for \code{function() { return(v) }}
}

