% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{EQ}
\alias{EQ}
\title{convinient functional version of ==}
\usage{
EQ(i)
}
\arguments{
\item{i}{the rhs argument to ==}
}
\value{
a function of arity 1 that retutns the result of comparing its argument with i
}
\description{
convinient functional version of ==
}

