% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%&&&\%}
\alias{\%&&&\%}
\title{(&&&) :: (b -> c) -> (b -> c') -> (b -> (c, c'))}
\usage{
f \%&&&\% g
}
\arguments{
\item{f}{The function whos return value is the first element of the resulting pair.}

\item{g}{The function whos return value is the second element of the resulting pair.}
}
\value{
A function of arity one that returns a pair. r(x) = (f(x), g(x))
}
\description{
(&&&) :: (b -> c) -> (b -> c') -> (b -> (c, c'))
}
\details{
This is the function instance's &&& of Arrow.
}
\references{
Haskell's Control.Arrow
}

