% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%$\%}
\alias{\%$\%}
\title{Funcion application}
\usage{
f \%$\% v
}
\arguments{
\item{f}{A function with arity 1}

\item{v}{The argunment to F}
}
\value{
The result of applying f to v
}
\description{
Funcion application
}
\details{
This relieves us from the burden of using ')' at the end.
}

