% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%cmap\%}
\alias{\%cmap\%}
\title{Apply a function to a dataframe's or matrix columns}
\usage{
f \%cmap\% d
}
\arguments{
\item{f}{The function to apply to each column of d.}

\item{d}{The dataframe or matrix.}
}
\value{
The result of applying f to the columns of d.
}
\description{
Apply a function to a dataframe's or matrix columns
}

