% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%o\%}
\alias{\%o\%}
\title{Function composition}
\usage{
f \%o\% g
}
\arguments{
\item{f}{A function with arity 1}

\item{g}{A function with arity 1}
}
\value{
A function with arity 1 resulting from composing f and g.
}
\description{
Function composition
}
\details{
\%o\% :: (a -> b) -> (c -> a) -> (c -> b)
}

