% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%select\%}
\alias{\%select\%}
\title{select according to a predicate}
\usage{
x \%select\% p
}
\arguments{
\item{x}{The collections of elements.}

\item{p}{(a -> Bool) The predicate that the returned elements should verify.}
}
\value{
The subset of elements of x that verify p
}
\description{
select according to a predicate
}

