% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{\%<-\%}
\alias{\%<-\%}
\title{persist a valut to storage}
\usage{
n \%<-\% value
}
\arguments{
\item{n}{The name where to bind the returned function}

\item{value}{The value to compute.}
}
\value{
A constant function. The first time this function is called, it will
try to read the persisted value from disk. If the file is not present, it
will compute and persist it.
}
\description{
persist a valut to storage
}

