% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GT}
\alias{GT}
\title{convinient functional version of >}
\usage{
GT(i)
}
\arguments{
\item{i}{the rhs argument to >}
}
\value{
a function of arity 1 that return the result of comparing its argument with i
}
\description{
convinient functional version of >
}

