% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lazy}
\alias{lazy}
\title{creates a lazy value}
\usage{
lazy(v = stop("v not supplied"))
}
\arguments{
\item{v}{the value that should be computed only when required}
}
\value{
a constant function (i.e. a function with no arguments). The first
time this function is called the computation v is performed and its value stored
in memory to be recalled upon succesive calls.
}
\description{
creates a lazy value
}
\details{
The value is computed only once and reused upon succesive calls.
}

