% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{LTE}
\alias{LTE}
\title{convinient functional version of <=}
\usage{
LTE(i)
}
\arguments{
\item{i}{the rhs argument to <=}
}
\value{
a function of arity 1 that returns the result of comparing its argument with i
}
\description{
convinient functional version of <=
}

