% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{selectBy}
\alias{selectBy}
\title{generate a function that selects according to a predicate}
\usage{
selectBy(p)
}
\arguments{
\item{p}{(a -> Bool) a predicate.}
}
\value{
A function of arity 1 that takes a collection and returns that collection's subset of elements that verify p
}
\description{
generate a function that selects according to a predicate
}
\seealso{
\%select\%
}

