% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{uncurry}
\alias{uncurry}
\title{Generates function that takes a pair as argument}
\usage{
uncurry(f)
}
\arguments{
\item{f}{A function of arity 2.}
}
\value{
A function that takes a pair as argument
}
\description{
Generates function that takes a pair as argument
}

